Move AppService and Cosmos to userSelection/services
