+++ src/client/src/containers/AppServiceModal/index.tsx
@@ -13,11 +13,11 @@ import buttonStyles from "../../css/buttonStyles.module.css";
 import { WIZARD_CONTENT_INTERNAL_NAMES, KEY_EVENTS, AZURE, EXTENSION_COMMANDS, SERVICE_KEYS, AzureResourceType } from "../../utils/constants";
 import styles from "./styles.module.css";
 import { AppState } from "../../store/combineReducers";
-import { ISelectedAppService } from "../../store/azureProfileData/appService/model";
-import { getAppServiceSelectionSelector } from "../../store/azureProfileData/appService/selector";
+import { ISelectedAppService } from "../../store/userSelection/services/appService/model";
+import { getAppServiceSelectionSelector } from "../../store/userSelection/services/appService/selector";
 import classNames from "classnames";
 import { useState } from "react";
-import { saveAppServiceSettingsAction } from "../../store/azureProfileData/appService/action";
+import { saveAppServiceSettingsAction } from "../../store/userSelection/services/appService/action";
 import { closeModalAction } from "../../store/navigation/modals/action";
 import { sendTelemetry } from "../../utils/extensionService/extensionService";
 import LocationSelection from "../../components/LocationSelection";
