+++ src/client/src/containers/CosmosDbModal/index.tsx
@@ -12,12 +12,12 @@ import buttonStyles from "../../css/buttonStyles.module.css";
 import { WIZARD_CONTENT_INTERNAL_NAMES, KEY_EVENTS, AZURE, EXTENSION_COMMANDS, SERVICE_KEYS, AzureResourceType } from "../../utils/constants";
 import styles from "./styles.module.css";
 import { AppState } from "../../store/combineReducers";
-import { ISelectedCosmosService } from "../../store/azureProfileData/cosmosDb/model";
-import { getCosmosDbSelectionSelector } from "../../store/azureProfileData/cosmosDb/selector";
+import { ISelectedCosmosService } from "../../store/userSelection/services/cosmosDb/model";
+import { getCosmosDbSelectionSelector } from "../../store/userSelection/services/cosmosDb/selector";
 import classNames from "classnames";
 import { useState } from "react";
 import { closeModalAction } from "../../store/navigation/modals/action";
-import { saveCosmosDbSettingsAction } from "../../store/azureProfileData/cosmosDb/action";
+import { saveCosmosDbSettingsAction } from "../../store/userSelection/services/cosmosDb/action";
 import { sendTelemetry } from "../../utils/extensionService/extensionService";
 import LocationSelection from "../../components/LocationSelection";
 import { ReactComponent as ArrowDown } from "../../assets/chevron.svg";
