+++ src/client/src/containers/PageAzureLogin/AzureSubscriptions/index.tsx
@@ -16,7 +16,7 @@ import {
   injectIntl
 } from "react-intl";
 import { AppState } from "../../../store/combineReducers";
-import { isAppServiceSelectedSelector } from "../../../store/azureProfileData/appService/selector";
+import { isAppServiceSelectedSelector } from "../../../store/userSelection/services/appService/selector";
 import messages from "./messages";
 import { setPageWizardPageAction, setDetailPageAction } from "../../../store/navigation/routes/action";
 import { isLoggedInSelector } from "../../../store/config/azure/selector";
