+++ src/client/src/store/userSelection/services/combineReducers.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from "redux";
-import cosmosDB from "../../azureProfileData/cosmosDb/reducer";
-import appService from "../../azureProfileData/appService/reducer";
+import cosmosDB from "./cosmosDb/reducer";
+import appService from "./appService/reducer";
 
 const serviceReducer = combineReducers({
   cosmosDB,
